refactor: separar validadores por rota e extrair validador compartilhado
