+++ src/utils/validators/__init__.py
@@ -1,0 +1,1 @@
+"""Pacote de validações separado por blueprint de rota."""
