+++ src/utils/validators/validator.py
@@ -1,0 +1,8 @@
+from flask import abort
+
+
+def require_fields(body: dict, *fields: str):
+    """Aborta com 400 se algum campo obrigatório estiver ausente."""
+    missing = [f for f in fields if f not in body or body[f] is None]
+    if missing:
+        abort(400, description=f"campos obrigatórios ausentes: {', '.join(missing)}")
